requirements.txt
openai
pygame
speechrecognition
whisper
pydub
simpleaudio
sounddevice
TTS
numpy
python-dotenv